requirements.txt
cryptography
httpx
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib
python-multipart
pytest
pytest-asyncio
httpx
qrcode
Pillow

psutil
